+++ python/subprojects/testbed/mlrl/testbed/experiments/output/evaluation/extractor_regression.py
@@ -6,15 +6,14 @@ sinks.
 """
 from typing import Any
 
-from mlrl.common.config.options import Options
-from mlrl.common.data.arrays import enforce_dense
-from mlrl.common.data.types import Float32
-
 from mlrl.testbed.experiments.output.data import OutputValue
 from mlrl.testbed.experiments.output.evaluation.measurements import Measurements
 from mlrl.testbed.experiments.output.evaluation.measures import Measure
 from mlrl.testbed.experiments.output.evaluation.measures_regression import REGRESSION_EVALUATION_MEASURES
 from mlrl.testbed.experiments.output.evaluation.writer import EvaluationDataExtractor
+
+from mlrl.util.arrays import enforce_dense
+from mlrl.util.options import Options
 
 
 class RegressionEvaluationDataExtractor(EvaluationDataExtractor):
@@ -24,7 +23,7 @@ class RegressionEvaluationDataExtractor(EvaluationDataExtractor):
 
     def _update_measurements(self, measurements: Measurements, index: int, ground_truth: Any, predictions: Any,
                              options: Options):
-        ground_truth = enforce_dense(ground_truth, order='C', dtype=Float32)
+        ground_truth = enforce_dense(ground_truth, order='C')
         evaluation_measures = OutputValue.filter_values(REGRESSION_EVALUATION_MEASURES, options)
 
         for evaluation_measure in evaluation_measures:
